Revert renaming of custom-class-name.native.js → custom-class-name-native.js and relevant import statement.
